Refactored adapter out of fragment into LocationSearchComponent.kt
